requirements.txt
fastapi
uvicorn[standard]
streamlit
fpdf2
python-multipart
requests
